basic-agent/src/agents/basic-agent/pyproject.toml
[project]
name = "basic-agent"
version = "0.1.0"
description = "Basic AgentCore agent — synchronous invoke, no Step Functions"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bedrock-agentcore>=1.2.0",
    "strands-agents>=1.22.0",
    "boto3>=1.34.0",
]
